halfway works
This is the error now:
Error creating ticket: invalid literal for int() with base 10: '139557435ee1012225'

We are mostly there. It picks up on the messages but just needs some minor tweaks.

Sorry for all the formatting changes - DJ we might need to get together on a standard for formatting. Every commit I made makes formatting issues across every file.